[DF][NFC] Adapt tests to new logic
