+++ bindings/experimental/distrdf/test/test_callable_generator.py
@@ -76,23 +76,23 @@ class ComputationGraphGeneratorTest(unittest.TestCase):
         n6 = node.Filter()  # noqa: avoid PEP8 F841
 
         # Generate and execute the mapper
-        generator = ComputationGraphGenerator.ComputationGraphGenerator(
-            node.proxied_node)
+        graph_dict = hn.generate_graph_dict()
         mapper_func = ComputationGraphGenerator.generate_computation_graph
-        triggerables = mapper_func(hn, t, 0)
-        nodes = generator.get_action_nodes()
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
 
-        reqd_order = [1, 3, 2, 2, 3, 2]
+        # Required order in the list of returned values (the nodes are stored
+        # in DFS order the first time they are appended to the graph)
+        reqd_order = [1, 2, 2, 3, 3, 2]
 
         self.assertEqual(t.ord_list, reqd_order)
-        self.assertListEqual(nodes, [n5.proxied_node, n4.proxied_node])
+        self.assertListEqual(nodes, [n4.proxied_node, n5.proxied_node])
         self.assertListEqual(triggerables, [t, t])
 
     def test_mapper_with_pruning(self):
         """
         A test case to check that the mapper works even in the case of
         pruning.
-
         """
         # A mock RDF object
         t = ComputationGraphGeneratorTest.Temp()
@@ -109,20 +109,216 @@ class ComputationGraphGeneratorTest(unittest.TestCase):
         n5 = n1.Count()
         n6 = node.Filter()  # noqa: avoid PEP8 F841
 
+        # Until here the graph would be:
+        # [1, 2, 2, 3, 3, 2]
+
         # Reason for pruning (change of reference)
         n5 = n1.Filter()  # noqa: avoid PEP8 F841
 
-        # Generate and execute the mapper
-        generator = ComputationGraphGenerator.ComputationGraphGenerator(
-            node.proxied_node)
-        # Prune first
-        generator.headnode.graph_prune()
-        mapper_func = ComputationGraphGenerator.generate_computation_graph
-        triggerables = mapper_func(hn, t, 0)
-        nodes = generator.get_action_nodes()
+        # After the change of reference, it becomes
+        # [1, 2, 2, 3, 2, 2]
+        # that is, the Filter is appended at the end of the list, it is fine
+        # because it holds a reference to the ID of the father.
 
-        reqd_order = [1, 2, 2, 2, 3, 2]
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        reqd_order = [1, 2, 2, 3, 2, 2]
 
         self.assertEqual(t.ord_list, reqd_order)
         self.assertListEqual(nodes, [n4.proxied_node])
+        # One occurrence of 't' per action node
         self.assertListEqual(triggerables, [t])
+
+    def test_dfs_graph_with_pruning_transformations(self):
+        """
+        Test case to check that transformation nodes with no children and
+        no user references get pruned.
+
+        """
+        # A mock RDF object
+        t = ComputationGraphGeneratorTest.Temp()
+
+        # Head node
+        hn = create_dummy_headnode(1)
+        hn.backend = ComputationGraphGeneratorTest.TestBackend()
+        node = Proxy.TransformationProxy(hn)
+
+        # Graph nodes
+        n1 = node.Define()
+        n2 = node.Filter()
+        n3 = n2.Filter()
+        n4 = n3.Count()  # noqa: avoid PEP8 F841
+        n5 = n1.Filter()  # noqa: avoid PEP8 F841
+        n6 = node.Filter()  # noqa: avoid PEP8 F841
+
+        # Transformation pruning, n5 was earlier a transformation node
+        n5 = n1.Count()  # noqa: avoid PEP8 F841
+
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        reqd_order = [1, 2, 2, 3, 2, 3]
+
+        self.assertEqual(t.ord_list, reqd_order)
+        self.assertListEqual(nodes, [n4.proxied_node, n5.proxied_node])
+        # One occurrence of 't' per action node
+        self.assertListEqual(triggerables, [t, t])
+
+    def test_dfs_graph_with_recursive_pruning(self):
+        """
+        Test case to check that nodes in a DistRDF graph with no user references
+        and no children get pruned recursively.
+        """
+        # A mock RDF object
+        t = ComputationGraphGeneratorTest.Temp()
+
+        # Head node
+        hn = create_dummy_headnode(1)
+        hn.backend = ComputationGraphGeneratorTest.TestBackend()
+        node = Proxy.TransformationProxy(hn)
+
+        # Graph nodes
+        n1 = node.Define()
+        n2 = node.Filter()
+        n3 = n2.Filter()
+        n4 = n3.Count()  # noqa: avoid PEP8 F841
+        n5 = n1.Filter()  # noqa: avoid PEP8 F841
+        n6 = node.Filter()  # noqa: avoid PEP8 F841
+
+        # Remove user references from n4, n3, n2
+        n4 = n3 = n2 = None  # noqa: avoid PEP8 F841
+
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        reqd_order = [1, 2, 2]
+
+        self.assertEqual(t.ord_list, reqd_order)
+        self.assertListEqual(nodes, [])
+        # One occurrence of 't' per action node
+        self.assertListEqual(triggerables, [])
+
+    def test_dfs_graph_with_parent_pruning(self):
+        """
+        Test case to check that parent nodes with no user references don't
+        get pruned.
+        """
+        # A mock RDF object
+        t = ComputationGraphGeneratorTest.Temp()
+
+        # Head node
+        hn = create_dummy_headnode(1)
+        hn.backend = ComputationGraphGeneratorTest.TestBackend()
+        node = Proxy.TransformationProxy(hn)
+
+        # Graph nodes
+        n1 = node.Define()
+        n2 = node.Filter()
+        n3 = n2.Filter()
+        n4 = n3.Count()  # noqa: avoid PEP8 F841
+        n5 = n1.Filter()  # noqa: avoid PEP8 F841
+        n6 = node.Filter()  # noqa: avoid PEP8 F841
+
+        # Remove references from n2 (which shouldn't affect the graph)
+        n2 = None
+
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        reqd_order = [1, 2, 2, 3, 2, 2]
+        # Removing references from n2 will not prune any node
+        # because n2 still has children
+
+        self.assertEqual(t.ord_list, reqd_order)
+        self.assertListEqual(nodes, [n4.proxied_node])
+        # One occurrence of 't' per action node
+        self.assertListEqual(triggerables, [t])
+
+    def test_dfs_graph_with_computed_values_pruning(self):
+        """
+        Test case to check that computed values in action nodes get
+        pruned.
+        """
+        # A mock RDF object
+        t = ComputationGraphGeneratorTest.Temp()
+
+        # Head node
+        hn = create_dummy_headnode(1)
+        hn.backend = ComputationGraphGeneratorTest.TestBackend()
+        node = Proxy.TransformationProxy(hn)
+
+        # Graph nodes
+        n1 = node.Define()
+        n2 = node.Filter()
+        n3 = n2.Filter()
+        n4 = n3.Count()  # noqa: avoid PEP8 F841
+        n5 = n1.Filter()
+        n6 = n5.Count()
+        n7 = node.Filter()
+
+        # This is to make sure action nodes with
+        # already computed values are pruned.
+        n6.proxied_node.value = 1
+
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        # The node 'n6' will be pruned. Hence,
+        # there's only one '3' in this list.
+        reqd_order = [1, 2, 2, 3, 2, 2]
+
+        self.assertEqual(t.ord_list, reqd_order)
+        self.assertListEqual(nodes, [n4.proxied_node])
+        # One occurrence of 't' per action node
+        self.assertListEqual(triggerables, [t])
+
+    def test_dfs_graph_without_pruning(self):
+        """
+        Test case to check that node pruning does not occur if every node either
+        has children or some user references.
+
+        """
+        # A mock RDF object
+        t = ComputationGraphGeneratorTest.Temp()
+
+        # Head node
+        hn = create_dummy_headnode(1)
+        hn.backend = ComputationGraphGeneratorTest.TestBackend()
+        node = Proxy.TransformationProxy(hn)
+
+        # Graph nodes
+        n1 = node.Define()
+        n2 = node.Filter()
+        n3 = n2.Filter()
+        n4 = n3.Count()  # noqa: avoid PEP8 F841
+        n5 = n1.Count()  # noqa: avoid PEP8 F841
+        n6 = node.Filter()  # noqa: avoid PEP8 F841
+
+        # Generate and execute the mapper
+        graph_dict = hn.generate_graph_dict()
+        mapper_func = ComputationGraphGenerator.generate_computation_graph
+        triggerables = mapper_func(graph_dict, t, 0)
+        nodes = hn.get_action_nodes()
+
+        reqd_order = [1, 2, 2, 3, 3, 2]
+
+        self.assertEqual(t.ord_list, reqd_order)
+        self.assertListEqual(nodes, [n4.proxied_node, n5.proxied_node])
+        # One occurrence of 't' per action node
+        self.assertListEqual(triggerables, [t, t])
