+++ bindings/experimental/distrdf/test/test_proxy.py
@@ -3,7 +3,7 @@ import unittest
 from DistRDF import Node
 from DistRDF import Proxy
 from DistRDF.Backends import Base
-
+from DistRDF.HeadNode import HeadNode
 
 class ProxyInitTest(unittest.TestCase):
     """Proxy abstract class cannot be instantiated."""
@@ -25,7 +25,7 @@ class TypeReturnTest(unittest.TestCase):
         TransformationProxy object is of type `DistRDF.TransformationProxy` and
         wraps a node object.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.TransformationProxy(node)
         self.assertIsInstance(proxy, Proxy.TransformationProxy)
@@ -36,7 +36,7 @@ class TypeReturnTest(unittest.TestCase):
         ActionProxy object is of type `DistRDF.ActionProxy` and
         wraps a node object.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.ActionProxy(node)
         self.assertIsInstance(proxy, Proxy.ActionProxy)
@@ -71,7 +71,7 @@ class AttrReadTest(unittest.TestCase):
 
     def test_attr_simple_action(self):
         """ActionProxy object reads the right input attribute."""
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.ActionProxy(node)
         func = proxy.attr
@@ -84,7 +84,7 @@ class AttrReadTest(unittest.TestCase):
         TransformationProxy object reads the right input attributes,
         returning the methods of the proxied node.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = AttrReadTest.TestBackend()
         proxy = Proxy.TransformationProxy(node)
 
@@ -106,17 +106,17 @@ class AttrReadTest(unittest.TestCase):
         When a node attribute is called on a TransformationProxy object, it
         correctly returns the attribute of the proxied node.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = AttrReadTest.TestBackend()
         proxy = Proxy.TransformationProxy(node)
 
         node_attributes = [
             "get_head",
             "operation",
-            "children",
+            "nchildren",
             "_new_op_name",
             "value",
-            "pyroot_node",
+            "rdf_node",
             "has_user_references"
         ]
 
@@ -129,7 +129,7 @@ class AttrReadTest(unittest.TestCase):
         When a non-defined Node class attribute is called on a
         TransformationProxy object, it raises an AttributeError.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.TransformationProxy(node)
         with self.assertRaises(AttributeError):
@@ -142,7 +142,7 @@ class AttrReadTest(unittest.TestCase):
         `__del__` method switches the node attribute `has_user_references` from
         `True` to `False`.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.TransformationProxy(node)
         self.assertTrue(node.has_user_references)
@@ -155,7 +155,7 @@ class AttrReadTest(unittest.TestCase):
         function call.
         """
         t = AttrReadTest.Temp()
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         node.value = t
         proxy = Proxy.ActionProxy(node)
@@ -194,7 +194,7 @@ class GetValueTests(unittest.TestCase):
         method in Proxy when the current action node
         already houses a value.
         """
-        node = Node.Node(None, None)
+        node = HeadNode()
         node.backend = None
         proxy = Proxy.ActionProxy(node)
         node.value = 5
